requirements.txt
fastapi[standard]==0.115.14
Pillow==11.3.0
filetype==1.2.0
torch==2.7.1
optimum[onnxruntime]==1.26.1